Move route components into auth feature folder
