+++ frontend/src/features/auth/components/routes/ProtectedRoute.tsx
@@ -1,0 +1,15 @@
+import { Navigate } from "react-router-dom";
+
+import useIsAuthenticated from "../../useIsAuthenticated";
+
+function ProtectedRoute({ children }) {
+	const isAuthenticated = useIsAuthenticated();
+
+	if (isAuthenticated === null) {
+		return <div>Loading ...</div>;
+	}
+
+	return isAuthenticated ? children : <Navigate to={"/login"} />;
+}
+
+export default ProtectedRoute;
